Update from upstream

+++ trigger/conf/global_settings.py
@@ -24,7 +24,13 @@ USE_GPG_AUTH = False
 USER_HOME = os.getenv('HOME')
 TACACSRC = os.getenv('TACACSRC', os.path.join(USER_HOME, '.tacacsrc'))
 TACACSRC_KEYFILE = os.getenv('TACACSRC_KEYFILE', os.path.join(PREFIX, '.tackf'))
-TACACSRC_PASSPHRASE = '' # NYI
+
+# If set, use the TACACSRC_PASSPHRASE, otherwise default to TACACSRC_KEYFILE
+TACACSRC_USE_PASSPHRASE = False
+
+# Use this passphrase to encrypt credentials.CHANGE THIS IN YOUR FILE BEFORE
+# USING THIS IN YOUR ENVIRONMENT.
+TACACSRC_PASSPHRASE = ''
 
 # Default login realm to store user credentials (username, password) for
 # general use within the .tacacsrc
@@ -63,6 +69,7 @@ SUPPORTED_VENDORS = (
     'a10',
     'arista',
     'aruba',
+    'avocent',
     'brocade',
     'cisco',
     'citrix',
@@ -74,6 +81,7 @@ SUPPORTED_VENDORS = (
     'mrv',
     'netscreen',
     'paloalto',
+    'pica8',
 )
 VALID_VENDORS = SUPPORTED_VENDORS # For backwards compatibility
 
@@ -82,11 +90,12 @@ VALID_VENDORS = SUPPORTED_VENDORS # For backwards compatibility
 # Trigger.
 #
 # If your internal definition differs from the UPPERCASED ones specified below
-# (which they probably do, customize them here.
+# (which they probably do), customize them here.
 VENDOR_MAP = {
     'A10 NETWORKS': 'a10',
     'ARISTA NETWORKS': 'arista',
     'ARUBA NETWORKS': 'aruba',
+    'AVOCENT': 'avocent',
     'BROCADE': 'brocade',
     'CISCO SYSTEMS': 'cisco',
     'CITRIX': 'citrix',
@@ -97,6 +106,7 @@ VENDOR_MAP = {
     'JUNIPER': 'juniper',
     'MRV': 'mrv',
     'NETSCREEN TECHNOLOGIES': 'netscreen',
+    'PICA8': 'pica8',
 }
 
 # A dictionary keyed by manufacturer name containing a list of the device types
@@ -105,22 +115,30 @@ SUPPORTED_PLATFORMS = {
     'a10': ['SWITCH'],
     'arista': ['SWITCH'],                         # Your "Cloud" network vendor
     'aruba': ['SWITCH'],                          # Aruba Wi-Fi controllers
+    'avocent': ['CONSOLE'],
     'brocade': ['ROUTER', 'SWITCH'],
-    'cisco': ['ROUTER', 'SWITCH'],
+    'cisco': ['ROUTER', 'SWITCH', 'FIREWALL'],
     'citrix': ['SWITCH'],                         # Assumed to be NetScalers
     'dell': ['SWITCH'],
-    'f5': ['LOAD BALANCING', 'SWITCH'],
+    'f5': ['LOAD_BALANCER', 'SWITCH'],
     'force10': ['ROUTER', 'SWITCH'],
     'foundry': ['ROUTER', 'SWITCH'],
     'juniper': ['FIREWALL', 'ROUTER', 'SWITCH'],  # Any devices running Junos
-    'mrv': ['CONSOLE SERVER', 'SWITCH'],
+    'mrv': ['CONSOLE', 'SWITCH'],
     'netscreen': ['FIREWALL'],                    # Pre-Juniper NetScreens
     'paloalto': ['FIREWALL'],
+    'pica8': ['ROUTER', 'SWITCH'],
 }
 
 # The tuple of support device types
-SUPPORTED_TYPES = ('CONSOLE SERVER', 'FIREWALL', 'DWDM', 'LOAD BALANCING',
-                   'ROUTER', 'SWITCH')
+SUPPORTED_TYPES = (
+    'CONSOLE',
+    'DWDM',
+    'FIREWALL',
+    'LOAD_BALANCER',
+    'ROUTER',
+    'SWITCH'
+)
 
 # A mapping of of vendor names to the default device type for each in the
 # event that a device object is created and the deviceType attribute isn't set
@@ -129,22 +147,29 @@ DEFAULT_TYPES = {
     'a10': 'SWITCH',
     'arista': 'SWITCH',
     'aruba': 'SWITCH',
+    'avocent': 'CONSOLE',
     'brocade': 'SWITCH',
     'citrix': 'SWITCH',
     'cisco': 'ROUTER',
     'dell': 'SWITCH',
-    'f5': 'LOAD BALANCING',
+    'f5': 'LOAD_BALANCER',
     'force10': 'ROUTER',
     'foundry': 'SWITCH',
     'juniper': 'ROUTER',
-    'mrv': 'CONSOLE SERVER',
+    'mrv': 'CONSOLE',
     'netscreen': 'FIREWALL',
     'paloalto': 'FIREWALL',
+    'pica8': 'SWITCH',
 }
 
 # When a vendor is not explicitly defined within `DEFAULT_TYPES`, fallback to
 # this type.
 FALLBACK_TYPE = 'ROUTER'
+
+
+# When a manufacturer/vendor is not explicitly defined, fallback to to this
+# value.
+FALLBACK_MANUFACTURER = 'UNKNOWN'
 
 #===============================
 # Twister
@@ -159,6 +184,15 @@ TELNET_TIMEOUT  = 60
 
 # Whether or not to allow telnet fallback
 TELNET_ENABLED = True
+
+# Default ports for SSH
+SSH_PORT = 22
+
+# Default port for Telnet
+TELNET_PORT = 23
+
+# The preferred order in which SSH authentication methods are tried.
+SSH_AUTHENTICATION_ORDER = ['password', 'keyboard-interactive', 'publickey']
 
 # A mapping of vendors to the types of devices for that vendor for which you
 # would like to disable interactive (pty) SSH sessions, such as when using
@@ -187,6 +221,22 @@ IOSLIKE_VENDORS = (
     'force10',
     'foundry',
 )
+
+# Prompts sent by devices that indicate the device is awaiting user
+# confirmation when interacting with the device. If a continue prompt is
+# detected, Trigger will temporarily set this value to the prompt and send
+# along the next command (for example if you're expecting such a prompt and you
+# want to send along "yes"). These should be as specific as possible because we
+# want to make sure bad things don't happen.
+CONTINUE_PROMPTS = [
+    'continue?',
+    'proceed?',
+    '(y/n):',
+    '[y/n]:',
+    '[confirm]',
+    '[yes/no]: ',
+    'overwrite file [startup-config] ?[yes/press any key for no]....'
+]
 
 # The file path where .gorc is expected to be found.
 GORC_FILE = '~/.gorc'
@@ -270,24 +320,26 @@ JUNIPER_FULL_COMMIT_FIELDS = {
 # Specially-defined, per-vendor prompt patterns. If a vendor isn't defined here,
 # try to use IOSLIKE_PROMPT_PAT or fallback to DEFAULT_PROMPT_PAT.
 PROMPT_PATTERNS = {
-    'aruba': r'\(\S+\)(?: \(\S+\))?\s?#', # ArubaOS 6.1
-    #'aruba': r'\S+(?: \(\S+\))?\s?#\s', # ArubaOS 6.2
+    'aruba': r'\(\S+\)(?: \(\S+\))?\s?#$', # ArubaOS 6.1
+    #'aruba': r'\S+(?: \(\S+\))?\s?#\s$', # ArubaOS 6.2
+    'avocent': r'\S+[#\$]|->\s?$',
     'citrix': r'\sDone\n$',
     'f5': r'.*\(tmos\).*?#\s{1,2}\r?$',
     'juniper': r'\S+\@\S+(?:\>|#)\s$',
     'mrv': r'\r\n?.*(?:\:\d{1})?\s\>\>?$',
     'netscreen': r'(\w+?:|)[\w().-]*\(?([\w.-])?\)?\s*->\s*$',
-    'paloalto': r'\r\n\S+(?:\>|#)\s?',
+    'paloalto': r'\r\n\S+(?:\>|#)\s?$',
+    'pica8': r'\S+(?:\>|#)\s?$',
 }
 
 # When a pattern is not explicitly defined for a vendor, this is what we'll try
 # next (since most vendors are in fact IOS-like)
-IOSLIKE_PROMPT_PAT = r'\S+(\(config(-[a-z:1-9]+)?\))?#'
-IOSLIKE_ENABLE_PAT = IOSLIKE_PROMPT_PAT[:-1] + '>'
+IOSLIKE_PROMPT_PAT = r'\S+(\(config(-[a-z:1-9]+)?\))?#\s?$'
+IOSLIKE_ENABLE_PAT = r'\S+(\(config(-[a-z:1-9]+)?\))?>\s?$'
 
 # Generic prompt to match most vendors. It assumes that you'll be greeted with
 # a "#" prompt.
-DEFAULT_PROMPT_PAT = r'\S+#'
+DEFAULT_PROMPT_PAT = r'\S+#\s?$'
 
 #===============================
 # Bounce Windows/Change Mgmt
@@ -419,7 +471,7 @@ BULK_MAX_HITS_DEFAULT = 1
 def _stage_acls(acls, log=None, sanitize_acl=False):
     """stage the new ACL files for load_acl"""
 
-    import os
+    import os, shutil
     from trigger.acl import parse as acl_parse
 
     acl_contents = []
@@ -432,9 +484,7 @@ def _stage_acls(acls, log=None, sanitize_acl=False):
         source = FIREWALL_DIR + '/%s' % acl
         dest = TFTPROOT_DIR + '/%s.%s' % (acl, nonce)
 
-        try:
-            os.stat(dest)
-        except OSError:
+        if not os.path.exists(dest):
             try:
                 shutil.copyfile(source, dest)
             except:
